Alert when sequencer is stuck
Algorithm:
Nonce gap=0 or sequencer_nonce nonce doesn't change, means no txn's going through, we're good.
If nonce_gap > 0 but changing and sequencer nonce changing: txn's are going through, we're good.
If nonce_gap > 0 and it's not changing and sequencer nonce isn't changing, we're stuck.
Alert will be sent to slack if we're stuck for over 5mins.